+++ src/app/plugins/plugin.module.ts
@@ -2,8 +2,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CreatePluginComponent} from './components/create-plugin/create-plugin.component';
 import { PluginsComponent} from './components/my-plugins/plugins.component';
-import { EditPluginComponent} from './components/edit-plugin/edit-plugin.component';
-import { PluginCardComponent} from './components/plugin-card/plugin-card.component';
 import {MatFormFieldModule} from '@angular/material/form-field';
 import {ReactiveFormsModule} from '@angular/forms';
 import {MaterialFileInputModule} from 'ngx-material-file-input';
@@ -11,9 +9,10 @@ import {MatIconModule} from '@angular/material/icon';
 import {MatProgressBarModule} from '@angular/material/progress-bar';
 import {RouterModule} from '@angular/router';
 import {MatButtonToggleModule} from '@angular/material/button-toggle';
-import {SharedModule} from '../shared/shared.module';
 import {MatInputModule} from '@angular/material/input';
 import {PluginService} from './services/plugin.service';
+import {MatTooltipModule} from '@angular/material/tooltip';
+import {EditPluginComponent} from './components/edit-plugin/edit-plugin.component';
 
 
 
@@ -22,8 +21,7 @@ import {PluginService} from './services/plugin.service';
     declarations: [
         CreatePluginComponent,
         PluginsComponent,
-        EditPluginComponent,
-        PluginCardComponent
+        EditPluginComponent
     ],
     imports: [
         CommonModule,
@@ -34,11 +32,11 @@ import {PluginService} from './services/plugin.service';
         MatProgressBarModule,
         RouterModule,
         MatButtonToggleModule,
-        SharedModule,
-        MatInputModule
+        MatInputModule,
+        MatTooltipModule
     ],
     exports: [
-        PluginCardComponent
+        CreatePluginComponent
     ],
     providers: [
         PluginService